Also inject env vars into build
