+++ packages/storybook-builder-vite/vite-server.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const {stringifyEnvs} = require('@storybook/core-common')
+const {stringifyProcessEnvs} = require('./envs');
 const { optimizeDeps } = require('./optimizeDeps');
 const { createServer } = require('vite');
 const { pluginConfig } = require('./vite-config');
@@ -46,23 +46,4 @@ module.exports.createViteServer = async function createViteServer(
 };
 
 
-/**
- * Customized version of stringifyProcessEnvs from @storybook/core-common which
- * uses import.meta.env instead of process.env
- */
- function stringifyProcessEnvs(raw) {
-    const envs = Object.entries(raw).reduce(
-        (acc, [key, value]) => {
-          acc[`import.meta.env.${key}`] = JSON.stringify(value);
-          return acc;
-        },
-        {
-          // Default fallback
-          'process.env.XSTORYBOOK_EXAMPLE_APP': '""',
-        }
-      );
-    // support destructuring like
-  // const { foo } = import.meta.env;
-  envs['import.meta.env'] = JSON.stringify(stringifyEnvs(raw));
-  return envs;
-}
+
